Add tests for translations. Save translations with the correct format for all entities
