+++ v4/Models/Survey.php
@@ -3,7 +3,9 @@
 namespace v4\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Validation\Rule;
 use Ushahidi\App\Repository\FormRepository;
+use Ushahidi\App\Validator\LegacyValidator;
 use Ushahidi\Core\Entity\Permission;
 use Ushahidi\Core\Tool\Permissions\InteractsWithFormPermissions;
 
@@ -114,8 +116,124 @@ class Survey extends Model
             ->where('form_stages.task_is_internal_only', '=', '0');
     }
 
-
-
+    protected static function getRules() {
+        return [
+            'name' => [
+                'required',
+                'min:2',
+                'max:255',
+                'regex:' . LegacyValidator::REGEX_STANDARD_TEXT
+            ],
+            'description' => [
+                'string',
+                'nullable'
+            ],
+            //@TODO find out where this color validator is implemented
+            //[['color']],
+            'color' => [
+                'string',
+                'nullable'
+            ],
+            'disabled' => [
+                'boolean'
+            ],
+            'hide_author' => [
+                'boolean'
+            ],
+            'hide_location' => [
+                'boolean'
+            ],
+            'hide_time' => [
+                'boolean'
+            ],
+            // @FIXME: disabled targeted survey creation for v4 forms, need to check
+            'targeted_survey' => [
+                Rule::in([false]),
+            ],
+            'tasks.*.label' => [
+                'required',
+                'regex:' . LegacyValidator::REGEX_STANDARD_TEXT
+            ],
+            'tasks.*.type' => [
+                Rule::in(['post', 'task'])
+            ],
+            'tasks.*.priority' => [
+                'numeric',
+            ],
+            'tasks.*.icon' => [
+                'alpha',
+            ],
+            'tasks.*.fields.*.label' => [
+                'required',
+                'max:150'
+            ],
+            'tasks.*.fields.*.key' => [
+                'max:150',
+                'alpha_dash'
+                // @TODO: add this validation for keys
+                //[[$this->repo, 'isKeyAvailable'], [':value']]
+            ],
+            'tasks.*.fields.*.input' => [
+                'required',
+                Rule::in([
+                    'text',
+                    'textarea',
+                    'select',
+                    'radio',
+                    'checkbox',
+                    'checkboxes',
+                    'date',
+                    'datetime',
+                    'location',
+                    'number',
+                    'relation',
+                    'upload',
+                    'video',
+                    'markdown',
+                    'tags',
+                ])
+            ],
+            'tasks.*.fields.*.type' => [
+                'required',
+                Rule::in([
+                    'decimal',
+                    'int',
+                    'geometry',
+                    'text',
+                    'varchar',
+                    'markdown',
+                    'point',
+                    'datetime',
+                    'link',
+                    'relation',
+                    'media',
+                    'title',
+                    'description',
+                    'tags',
+                ])
+                // @TODO: add this validation for duplicates in type?
+                //[[$this, 'checkForDuplicates'], [':validation', ':value']],
+            ],
+            'tasks.*.fields.*.type' => [
+                'boolean'
+            ],
+            'tasks.*.fields.*.priority' => [
+                'numeric',
+            ],
+            'tasks.*.fields.*.cardinality' => [
+                'numeric',
+            ],
+            'tasks.*.fields.*.response_private' => [
+                'boolean'
+                // @TODO add this custom validator for canMakePrivate
+                // [[$this, 'canMakePrivate'], [':value', $type]]
+            ]
+            // @NOTE: checkPostTypeLimit is not used here.
+            // Before merge, validate with Angela if we
+            // should be removing that arbitrary limit since it's pretty rare
+            // for it to be needed
+        ];
+    }
 
     /**
      * Get the survey's translation.
